Change banana spawner to be in its own game object
Also fixed bug of GameOver method not executing.
