+++ EscapeTheZoo/Assets/Scenes/CatchTheBananasScene.unity
@@ -721,6 +721,7 @@ RectTransform:
   - {fileID: 1485588735}
   - {fileID: 537222564}
   - {fileID: 25672683}
+  - {fileID: 1591303997}
   m_Father: {fileID: 0}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 2, z: 0}
@@ -1245,6 +1246,56 @@ MonoBehaviour:
   leftKey: 97
   rightKey: 100
   upKey: 119
+--- !u!1 &1591303996
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1591303997}
+  - component: {fileID: 1591303998}
+  m_Layer: 5
+  m_Name: BananaSpawner
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!224 &1591303997
+RectTransform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1591303996}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 629567254}
+  m_RootOrder: 4
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchoredPosition: {x: 0, y: 0}
+  m_SizeDelta: {x: 100, y: 100}
+  m_Pivot: {x: 0.5, y: 0.5}
+--- !u!114 &1591303998
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1591303996}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b98caae40049ef64abba52f5bbc98302, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  bananaPrefab: {fileID: 2035329849}
 --- !u!1001 &1611784735
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1265,6 +1316,18 @@ PrefabInstance:
       value: 3
       objectReference: {fileID: 0}
     - target: {fileID: 5214410937759287447, guid: e4ac42e8717250644b393e938eed36bb, type: 3}
+      propertyPath: m_LocalScale.x
+      value: 0.6
+      objectReference: {fileID: 0}
+    - target: {fileID: 5214410937759287447, guid: e4ac42e8717250644b393e938eed36bb, type: 3}
+      propertyPath: m_LocalScale.y
+      value: 0.6
+      objectReference: {fileID: 0}
+    - target: {fileID: 5214410937759287447, guid: e4ac42e8717250644b393e938eed36bb, type: 3}
+      propertyPath: m_LocalScale.z
+      value: 0.6
+      objectReference: {fileID: 0}
+    - target: {fileID: 5214410937759287447, guid: e4ac42e8717250644b393e938eed36bb, type: 3}
       propertyPath: m_LocalPosition.x
       value: 0
       objectReference: {fileID: 0}
@@ -1274,7 +1337,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5214410937759287447, guid: e4ac42e8717250644b393e938eed36bb, type: 3}
       propertyPath: m_LocalPosition.z
-      value: -9719.999
+      value: -9720
       objectReference: {fileID: 0}
     - target: {fileID: 5214410937759287447, guid: e4ac42e8717250644b393e938eed36bb, type: 3}
       propertyPath: m_LocalRotation.w
@@ -1469,7 +1532,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   gameOverCanvas: {fileID: 1024115994}
-  bananaPrefab: {fileID: 2035329849}
+  bananaSpawner: {fileID: 1591303998}
   player1ScoreText: {fileID: 2058154972}
   player2ScoreText: {fileID: 2104716267}
 --- !u!1 &2035329849 stripped
