+++ EscapeTheZoo/Assets/Scripts/CatchTheBananas/BananaSpawner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b98caae40049ef64abba52f5bbc98302
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
